A first attempt at re-designing the OpticalRawDigit to be less uboone-specific. - Inverted the inheritance so now FIFOChannel is derived from OpticalRawDigit. - OpticalRawDigit now has just a time, a channel, and the waveform. - A Frame() function which just returns zero is kept for now so that algorithms which expect multiple frames can still be used.  May want to deprecate this out in the future as algorithms are developed. - TimeSlice_t was made an unsigned long so it can hold the 64 bit NOvA-type time stamp used in the LBNE electronics.

+++ OpticalDetectorData/OpticalTypes.h
@@ -17,9 +17,10 @@ namespace optdata {
 
   // Other types, in case something changes later on:
   typedef unsigned int Channel_t;
-  typedef unsigned int TimeSlice_t;
+  typedef unsigned long TimeSlice_t;
   typedef unsigned int Frame_t;
 
+    //  uboone specific - move away
   // Use an enum for categories of ADC channels, triggers, etc.
   typedef enum _optical_category_t {
     kUndefined,
